# Bugfix: Characters in upper part of BMP would fail unescaping