**Description**: 
When escaping characters in the upper part (`]\uDFFF-\uFFFF[`) of the Basic Multilingual Plane (BMP), the `decodeUnicodeEscape` would fail to see that this was a single rune instead of complex UTF16 rune.
When checking for single rune we now check if correctly if the single rune is within the complete space of BMP `[\u0000-\uFFFF[` and that its not in `[\uD800-uDFFF]`.

Furthermore I've fixed a bug in ParseString, such that if a unescape error happens a `MalformedValueError` is returned. I've also added unit test for ParseString.

NB: To run the benchmarks in the docker container i changed the Dockerfile to use `golang:1.8` instead of `golang:1.6`.

**Benchmark before change**:
BenchmarkJsonParserLarge-4                    	  200000	     58070 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	 1000000	     10187 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      6662 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      6941 ns/op	     704 B/op	      13 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	 1000000	      9475 ns/op	     656 B/op	      12 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	       869 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       695 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	       972 ns/op	     256 B/op	       9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       829 ns/op	     240 B/op	       8 allocs/op

**Benchmark after change**:
BenchmarkJsonParserLarge-4                    	  100000	     65712 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserMedium-4                   	 1000000	     10498 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualMedium-4      	 1000000	      6568 ns/op	     112 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructMedium-4      	 1000000	      7012 ns/op	     704 B/op	      13 allocs/op
BenchmarkJsonParserObjectEachStructMedium-4   	 1000000	      9974 ns/op	     656 B/op	      12 allocs/op
BenchmarkJsonParserSmall-4                    	10000000	       917 ns/op	       0 B/op	       0 allocs/op
BenchmarkJsonParserEachKeyManualSmall-4       	10000000	       670 ns/op	      80 B/op	       2 allocs/op
BenchmarkJsonParserEachKeyStructSmall-4       	10000000	       955 ns/op	     256 B/op	       9 allocs/op
BenchmarkJsonParserObjectEachStructSmall-4    	10000000	       827 ns/op	     240 B/op	       8 allocs/op


For running benchmarks use:
```
go test -test.benchmem -bench JsonParser ./benchmark/ -benchtime 5s -v
# OR
make bench (runs inside docker)
```