@jstroem 2017-05-25
This fixes #104.
Would love to get it landed ASAP so any input would be fixed fast @buger.